HW10: move event to top-level domain/ module
