+++ hw12_13_14_15_calendar/domain/event.go
@@ -1,4 +1,4 @@
-package event
+package domain
 
 import (
 	"time"
