+++ hw12_13_14_15_calendar/internal/storage/sql/storage.go
@@ -2,7 +2,7 @@ package sqlstorage
 
 import (
 	"context"
-	"github.com/artembert/golang-pro-otus-hw/hw12_13_14_15_calendar/internal/domain/event"
+	"github.com/artembert/golang-pro-otus-hw/hw12_13_14_15_calendar/domain"
 	"github.com/artembert/golang-pro-otus-hw/hw12_13_14_15_calendar/internal/domain/storage"
 	"github.com/jackc/pgx/v4/pgxpool"
 )
@@ -40,24 +40,24 @@ func (s *Storage) Close(ctx context.Context) error {
 	return nil
 }
 
-func (s *Storage) CreateEvent(evt event.Event) error {
+func (s *Storage) CreateEvent(evt domain.Event) error {
 	return nil
 }
 
-func (s *Storage) DeleteEvent(evt event.Event) error {
+func (s *Storage) DeleteEvent(evt domain.Event) error {
 	return nil
 }
 
-func (s *Storage) UpdateEvent(evt event.Event) error {
+func (s *Storage) UpdateEvent(evt domain.Event) error {
 	return nil
 }
 
-func (s *Storage) GetEventByID(id string) (event.Event, error) {
-	return event.Event{}, nil
+func (s *Storage) GetEventByID(id string) (domain.Event, error) {
+	return domain.Event{}, nil
 }
 
-func (s *Storage) GetAllEvents() ([]event.Event, error) {
-	return make([]event.Event, 0), nil
+func (s *Storage) GetAllEvents() ([]domain.Event, error) {
+	return make([]domain.Event, 0), nil
 }
 
 // Compile-time check that Storage implements storage.Storage
